requirements.txt
streamlit
deepface==0.0.79
tensorflow-cpu==2.15.0
keras==2.15.0
opencv-python-headless
pillow
requests
google-api-python-client
pandas